requirements.txt
matplotlib==3.8.1
nltk==3.8.1
numpy==1.26.1
pandas==2.1.3
python-dotenv==1.0.0
streamlit==1.28.1
textblob==0.17.1
transformers==4.35.0
tweepy==4.14.0
wordcloud==1.9.2
